+++ components/RideLayout.tsx
@@ -1,8 +1,9 @@
 import { icons } from "@/constants";
 import { useRouter } from "expo-router";
-import { ReactNode } from "react";
+import { ReactNode, useRef } from "react";
 import { Image, Text, TouchableOpacity, View } from "react-native";
 import { GestureHandlerRootView } from "react-native-gesture-handler";
+import BottomSheet, { BottomSheetScrollView } from "@gorhom/bottom-sheet";
 import Map from "./Map";
 
 const RideLayout = ({
@@ -13,6 +14,9 @@ const RideLayout = ({
   children: ReactNode;
 }) => {
   const router = useRouter();
+
+  const bottomSheetRef = useRef<BottomSheet>(null);
+
   return (
     <GestureHandlerRootView>
       <View className="flex-1 bg-white">
@@ -33,6 +37,11 @@ const RideLayout = ({
           </View>
           <Map />
         </View>
+        <BottomSheet ref={bottomSheetRef} snapPoints={["40%", "85%"]} index={0}>
+          <BottomSheetScrollView style={{ flex: 1, padding: 20 }}>
+            {children}
+          </BottomSheetScrollView>
+        </BottomSheet>
       </View>
     </GestureHandlerRootView>
   );