+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- added bottom-sheet
 - ride layout component implemenation
 - find-ride route
 - fide ride component